+++ app/src/main/java/com/cscmobi/habittrackingandroid/thanhlv/database/Dao.kt
@@ -23,13 +23,13 @@ interface Dao {
     fun getAllTask(): Flow<List<Task>>
 
     @Query("SELECT * FROM task WHERE id IN (:tasksId)")
-    fun loadAllTaskByIds(tasksId: IntArray): Flow<List<Task>>
+    fun loadAllTaskByIds(tasksId: LongArray): Flow<List<Task>>
 
     @Query("SELECT * FROM task WHERE name LIKE :name")
     fun findTaskByName(name: String): Flow<Task>
 
     @Query("SELECT * FROM task WHERE id=:id")
-    fun findTaskById(id: Int): Flow<Task>
+    fun findTaskById(id: Long): Flow<Task>
 
     @Insert
     suspend fun insertAllTask(vararg users: Task)
@@ -122,7 +122,7 @@ interface Dao {
 
     @Query("UPDATE history SET taskInDay = :newTaskInDay WHERE id = :id")
 
-    suspend fun deleteTaskinHistory(id: Int, newTaskInDay: List<TaskInDay>)
+    suspend fun deleteTaskinHistory(id: Long, newTaskInDay: List<TaskInDay>)
 
     @Query("SELECT * FROM history WHERE date >= :startDate")
     fun getHistoryWithDate(startDate: Long): Flow<List<History>>
