+++ app/src/main/java/com/cscmobi/habittrackingandroid/thanhlv/model/History.kt
@@ -12,7 +12,7 @@ import org.checkerframework.common.aliasing.qual.Unique
 @Serializable
 @Entity(indices = [Index(value = ["date"], unique = true)])
 data class History(
-    @PrimaryKey(true) var id: Int = 0,
+    @PrimaryKey(true) var id: Long = 0,
     @ColumnInfo(name = "date")  var date: Long? = 0,
     @ColumnInfo(name = "taskInDay") var taskInDay: List<TaskInDay> = listOf(),
     @ColumnInfo(name = "progressDay") var progressDay: Int = 0 //3*100/5 -> int
